requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
idna==2.8
IMDbPY==6.6
kiwisolver==1.1.0
lxml==4.3.3
matplotlib==3.0.3
nltk==3.4.1
numpy==1.16.3
Pillow==6.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
requests==2.21.0
six==1.12.0
soupsieve==1.9.1
SQLAlchemy==1.3.3
stemming==1.0.1
urllib3==1.24.3
wordcloud==1.5.0
